requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
sentence_transformers
langchain_huggingface
faiss-cpu
pandas
langchain-groq
duckduckgo_search==5.3.1b1
pymupdf
huggingface_hub
streamlit